+++ app/Auth/routes.js
@@ -4,12 +4,16 @@
 const Router = require('express').Router()
 
 const controller = require('./controller')
-const { facebook } = require('./strategies')
+const {
+  github,
+  facebook
+} = require('./strategies')
 
 Router.get('/', (request, response) => {
   response.json({})
 })
 
+github(Router, controller)
 facebook(Router, controller)
 
 Router.get('/error', controller.fail)